move to new package rss2json in model package
